% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_has_op}
\alias{check_has_op}
\title{Check if 1Password CLI app op is available/working}
\usage{
check_has_op(version = "0.4")
}
\arguments{
\item{version}{version of \code{op} to test for the presence of}
}
\value{
\code{NULL}, invisibly. This function is called to check for the presence
of \code{op} and will fail if it is not found with the specified version.
}
\description{
Check if 1Password CLI app op is available/working
}
