% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1password.R
\name{find_item}
\alias{find_item}
\title{Find a 1Password item matching a name}
\usage{
find_item(name)
}
\arguments{
\item{name}{a name to seach for within 1Password item titles}
}
\value{
a character vector of 1Password item titles matching `name`
}
\description{
Find a 1Password item matching a name
}
\examples{
\dontrun{
find_item("Twitter")}
}
