% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_item}
\alias{get_item}
\title{Retrieve a 1Password item by name}
\usage{
get_item(name)
}
\arguments{
\item{name}{name of item to retrieve. See Details.}
}
\value{
the input username for the given item
}
\description{
Retrieve a 1Password item by name
}
\details{
If \code{name} is not unique the shell process will return the available
unique hashes which \emph{are} unique and can be used as \code{name} arguments in a
further call.
}
\examples{
\dontrun{
get_item("Twitter")}
}
