% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_token_name}
\alias{get_token_name}
\title{Retrieve the environment variable name}
\usage{
get_token_name()
}
\description{
Idenfities the current environment variable name by
inspecting \code{getOption("OP_SUBDOMAIN")} which is set on
signin.
}
