% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opr-package.R
\docType{package}
\name{opr-package}
\alias{opr}
\alias{opr-package}
\title{opr: OpenPrescribing in R}
\description{
Provides functions to connect to and query the OpenPrescribing.net BigQuery database. Helps researchers understand the available data and write readable, reusable, and modular queries for analysing English prescribing data using 'dplyr' and 'dbplyr' instead of SQL.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bennettoxford/opr}
  \item \url{https://bennettoxford.github.io/opr/}
  \item Report bugs at \url{https://github.com/bennettoxford/opr/issues}
}

}
\author{
\strong{Maintainer}: Milan Wiedemann \email{milan.wiedemann@gmail.com} (\href{https://orcid.org/0000-0003-1991-282X}{ORCID})

Authors:
\itemize{
  \item Milan Wiedemann \email{milan.wiedemann@gmail.com} (\href{https://orcid.org/0000-0003-1991-282X}{ORCID})
}

}
\keyword{internal}
