% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalReg.R
\name{ordinalReg}
\alias{ordinalReg}
\title{Fits Semiparametric Regression Models for Ordinal Panel Count Data}
\usage{
ordinalReg(
  formula,
  data,
  method = c("ML", "Random", "EM"),
  se = c("NULL", "Bootstrap", "Fisher"),
  control = list()
)
}
\arguments{
\item{formula}{A formula object, with the response on the left of a
"~" operator, and the terms on the right. The response must be a
ordinal survival object as returned by function \code{OrdinalSurv}.}

\item{data}{A data.frame in which to interpret the variables named in
the \code{formula}. Three variables including subjects' id,
observation times, and ordinal observation of new events since last
observation time are required to feed into function
\code{OrdinalSurv} as response. At least one covariate variable is required.}

\item{method}{Fitting method. See \sQuote{Details}.}

\item{se}{Standard error estimation method. See \sQuote{Details}.}

\item{control}{A list of control parameters. See \sQuote{Details}.}
}
\value{
An object of S3 class \code{"ordinalReg"} representing the fit results.
}
\description{
Fits Semiparametric Regression Models for Ordinal Panel Count Data
}
