% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdinalSurv.R
\name{OrdinalSurv}
\alias{OrdinalSurv}
\title{Create a OrdinalSurv Object}
\usage{
OrdinalSurv(id, time, ord, lower = NULL, upper = NULL)
}
\arguments{
\item{id}{Observation subject's ID.}

\item{time}{Observation time.}

\item{ord}{Ordinal observation of event count}

\item{lower}{(Optional) lower cut points for ordinal observations}

\item{upper}{(Optional) upper cut points for ordinal observations}
}
\value{
An object of S3 class \code{"OrdinalSurv"}.
}
\description{
Create a OrdinalSurv Object
}
