% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{practice_settings}
\alias{practice_settings}
\title{Practice setting codes and labels}
\format{
A tibble with 24 rows and 2 columns:
\describe{
\item{setting}{Integer setting code}
\item{setting_label}{Descriptive label}
}
}
\usage{
practice_settings
}
\description{
Lookup table mapping practice \code{setting} codes to descriptive labels,
e.g. setting \code{4} is a standard GP practice. Used by \code{\link[=get_practices]{get_practices()}}
to add the \code{setting_label} column.
}
\examples{
practice_settings
}
\keyword{datasets}
