% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{setup_op}
\alias{setup_op}
\title{Attempt to signin to 1Password for the first time.}
\usage{
setup_op(URL = "my.1password.com", email = NULL, secret_key = NULL)
}
\arguments{
\item{URL}{domain used to signin to 1Password, e.g. \code{my.1password.com}}

\item{email}{your login email}

\item{secret_key}{your Secret Key}
}
\value{
if successful, the login token, invisibly.
}
\description{
This function is only required once, after which point a new folder will be
created by \code{op} in the user's home directory, storing the session tokens. See
Details.
}
\details{
Requests the 1Password Master Password. At no point is this saved to
even a temporary variable. This is only required for signin, after which a
session token will be used to communicate with 1Password. The session token
is validated on the server and lasts 30 minutes, after which you will again
need to sign in.

The session token is not the actual hash, but merely a reference to an
encrypted token in the user's home directory (in e.g. \code{~/.op/config}).

This sets an environment variable \code{OP_SESSION_X} (where \code{X} is the subdomain)
and an \link{option} \code{OP_SUBDOMAIN}.
}
