% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/op.R
\name{signin}
\alias{signin}
\title{Attempt to sign in to an authenticated op session}
\usage{
signin(subdomain = "my")
}
\arguments{
\item{subdomain}{domain for logging into 1Password. This can be found in the
URL you use to login via a web interface, e.g. for
\code{my.1password.com/signin} the subdomain is \code{my}, which is set as default.}
}
\value{
if successful, the login token, invisibly.
}
\description{
Attempt to sign in to an authenticated op session
}
\details{
Requests the 1Password Master Password. At no point is this saved to
even a temporary variable. This is only required for signin, after which a
session token will be used to communicate with 1Password. The session token
is validated on the server and lasts 30 minutes, after which you will again
need to sign in.

The session token is not the actual hash, but merely a reference to an
encrypted token in the user's home directory (in e.g. \code{~/.op/config}).

This sets an environment variable \code{OP_SESSION_X} (where \code{X} is the subdomain)
and an \link{option} \code{OP_SUBDOMAIN}.
}
