% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opr-package.R
\docType{package}
\name{opr-package}
\alias{opr}
\alias{opr-package}
\title{opr: Interact with the 1Password CLI tool 'op'}
\description{
Query and retrieve 1Password usernames and passwords
via the CLI tool 'op'.
}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (0000-0002-1404-5264)

}
\keyword{internal}
